qr_generator/qr_requirements.txt
qrcode==6.1
Pillow==6.2.2
fpdf==1.7.2